# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Admin Link Pool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Pool {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  createdAt: string;
  /** @nullable */
  linkCount?: number | null;
  /** @nullable */
  totalClicks?: number | null;
}

export interface PoolInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  color?: string;
}

export interface PoolUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  color?: string;
}

export interface Link {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  poolId?: number | null;
  /** @nullable */
  poolName?: string | null;
  /** @nullable */
  poolColor?: string | null;
  clicks?: number;
  isActive?: boolean;
  createdAt: string;
}

export interface LinkInput {
  /** @minLength 1 */
  title: string;
  url: string;
  description?: string;
  /** @nullable */
  poolId?: number | null;
  isActive?: boolean;
}

export interface LinkUpdate {
  /** @minLength 1 */
  title?: string;
  url?: string;
  description?: string;
  /** @nullable */
  poolId?: number | null;
  isActive?: boolean;
}

export interface StatsOverview {
  totalLinks: number;
  totalPools: number;
  totalClicks: number;
  activeLinks: number;
  recentLinks?: Link[];
}

export interface PoolStat {
  poolId: number;
  poolName: string;
  /** @nullable */
  poolColor?: string | null;
  linkCount: number;
  totalClicks: number;
}

export type ListLinksParams = {
/**
 * @nullable
 */
poolId?: number | null;
/**
 * @nullable
 */
search?: string | null;
};

export type GetTopLinksParams = {
/**
 * @nullable
 */
limit?: number | null;
};