# lib/api-zod/src/generated/types/link.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Admin Link Pool API
 * OpenAPI spec version: 0.1.0
 */

export interface Link {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  poolId?: number | null;
  /** @nullable */
  poolName?: string | null;
  /** @nullable */
  poolColor?: string | null;
  clicks?: number;
  isActive?: boolean;
  createdAt: string;
}