# lib/api-zod/src/generated/types/linkUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Admin Link Pool API
 * OpenAPI spec version: 0.1.0
 */

export interface LinkUpdate {
  /** @minLength 1 */
  title?: string;
  url?: string;
  description?: string;
  /** @nullable */
  poolId?: number | null;
  isActive?: boolean;
}