# lib/api-zod/src/generated/types/pool.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Admin Link Pool API
 * OpenAPI spec version: 0.1.0
 */

export interface Pool {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  createdAt: string;
  /** @nullable */
  linkCount?: number | null;
  /** @nullable */
  totalClicks?: number | null;
}