# lib/api-zod/src/generated/types/poolStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Admin Link Pool API
 * OpenAPI spec version: 0.1.0
 */

export interface PoolStat {
  poolId: number;
  poolName: string;
  /** @nullable */
  poolColor?: string | null;
  linkCount: number;
  totalClicks: number;
}