# lib/api-zod/src/generated/types/poolUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Admin Link Pool API
 * OpenAPI spec version: 0.1.0
 */

export interface PoolUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  color?: string;
}