# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Admin Link Pool API
 * OpenAPI spec version: 0.1.0
 */
import type { Link } from './link';

export interface StatsOverview {
  totalLinks: number;
  totalPools: number;
  totalClicks: number;
  activeLinks: number;
  recentLinks?: Link[];
}